[TensorExpr] add support for Reduction Ops

+++ torch/csrc/jit/tensorexpr/expr.h
@@ -181,7 +181,7 @@ class TORCH_API Buf : public ExprNode<Buf> {
     TORCH_CHECK(var);
   }
 
-  int ndim() const {
+  size_t ndim() const {
     return dims_.size();
   }
   const Expr* dim(int index) const {
